Improve speed and accuracy of EdgeAligner
Previously, the overlapping slices were subpixel-shifted to correspond
exactly with the reported stage positions. Now, we only do whole-pixel
shifts and add the remaining fractions back on to the computed offset.
This greatly improves the speed as the subpixel shift was one of the
slowest operations in the registration logic, as well as potentially
improving accuracy as the shift could potentially have degraded the signal.