Change LayerAligner strategy to only use original tiles
Before, LayerAligner was slicing regions out of the base mosaic for aligning
tiles on subsequent layers. Now it finds the nearest neighbor tile from the
base tile set based on stage positions, and aligns only the overlapping region
of those images. This should provide a cleaner signal to align on, as the
tiles in the assembled base mosaic have gone through a subpixel shift pass
which might introduce artifacts. It also means we don't need to keep the base
mosaic in memory anymore. Runtime seems to have degraded, but I haven't
formally measured it yet.