Make EdgeAligner linear model fit aware of "islands"
For samples with separated islands of cells, only run the linear model
fitting on the largest connected component of the induced spanning tree.
Then correct the relative positions of the remaining components based on
the model. This improves the large-scale tile positioning accuracy as well
as the accuracy of the linear model, therefore improving subsequent
layer alignment passes as well.