Perform alignment with multiple overlap window sizes
This makes the alignment much more robust against large variances in stage
positioning error, at the expense of some runtime. We start with the "native"
overlap window and double it until the window is greater than or equal to that
needed to see the maximum allowable shift.  Typically this only requires one
doubling so the performance is about half of what it was before this change.