Improve plot utility functions
Calls to modest_image have been refactored into a helper function. This
simplified the code and allowed for a more explicit warning message under
Python 3 where modest_image is not available. Additionally, the new function
now draws a background rectangle to make the colored graph edges as visible
as they are when the image is drawn (the light colors were not visible on
a white background).