Improve ergonomics of command line tool and some internal functionality (#16)
* Improve plot utility functions

Calls to modest_image have been refactored into a helper function. This
simplified the code and allowed for a more explicit warning message under
Python 3 where modest_image is not available. Additionally, the new function
now draws a background rectangle to make the colored graph edges as visible
as they are when the image is drawn (the light colors were not visible on
a white background).

* Improve flexibility of command line tool

* Add -o output path option
* Add -f output filename format option (also default is changed from
  previous hard-coded value)
* Cleanly deinitialize bioformats more reliably to avoid hangs
* Change from "scan" terminology to "cycle"

* Add debug and mode arguments to Mosaic.run

debug generates a color image where each tile is a different pure RGB color
and the tiles are combined by addition. This helps highlight alignment problems
by showing a color fringe around misaligned objects.

mode allows the non-default value 'return' which will return a list of images
without writing them to files.