Switch from javabridge + python-bioformats to pyjnius + our own copy of the JAR (#17)
* Switch to pyjnius for interfacing with Bioformats

* Upgrade versioneer to commit 6b555078bec6286dbbb46f8ae9c775597e162186

We need an as-yet-unreleased fix (cmdclass overrides).

* Switch to our own copy of the bioformats JAR

This involves overrding a few setuptools commands to first fetch the JAR
from the openmicroscopy site. Critically the JAR should always end up
included in sdists, so users installing from pypi will already have it
in their downloaded package. The JAR is retained locally as well so
developers won't have to keep re-downloading it either.