Various improvements to output quality (#21)
* Add support for increasing the overlap region

Also swap the sign of some calculations for consistency.

* Refactor intersection result values into a class

* Perform alignment with multiple overlap window sizes

This makes the alignment much more robust against large variances in stage
positioning error, at the expense of some runtime. We start with the "native"
overlap window and double it until the window is greater than or equal to that
needed to see the maximum allowable shift.  Typically this only requires one
doubling so the performance is about half of what it was before this change.

* Expose mosaic channel selection to command line tool

* Fix edge layout logic around padding and fractional offsets

* Build spanning tree with single source dijkstra path

This seems to do a better job than the previous greedy algorithm.

* Add alternate paste modes and default to linear blend