Add filepattern reader
This is a renamed and updated NistReader, which can now handle multi-channel
datasets. Either the image tiles can be multi-channel files, or they can be
single-channel files with a channel name/number encoded in the filename.

This change also adds undocumented support for alternative file readers to
the "ashlar" command line script. Input file paths can now be pipe-delimited
strings encoding the reader name, base path, and keyword arguments for the
reader class. For example, with the new filepattern reader:

    filepattern|/path/to/images|pattern=img_r{row:03}_c{col:03}.tif|overlap=0.1

Be sure to quote such strings to avoid the pipe characters being interpreted
by your shell.