Fix aligning multiple single-tile cycles
With one tile, the EdgeAligner linear model ended up with a transform
matrix of all zeros, which led the LayerAligner to make incorrect
predictions about the expected location of the tiles in the later cycles
and thus reject them too often. In this change we detect the degenerate
transform and reset it to the identity matrix.