Drop pyfftw in favor of scipy's pocketfft
Pyfftw is much faster than fftpack, enough that it was worthwhile for us to
deal with the difficulty of getting it installed. Now scipy has switched to
pocketfft which is fast enough and much easier to build (even included in scipy
binary wheels) so we'll use that instead. Pocketfft also seems to be easier
than pyfftw to use correctly and efficiently from multiple threads. Note that
while numpy also switched to pocketfft, they are using an older and apparently
slower version. Scikit-image has come to the same conclusion and now uses
scipy.fft as well instead of numpy.fft, so we don't even need to monkey-patch
anything.