Improve EdgeAligner robustness with small overlaps
This change drastically reduces the need to tune the max_shift parameter.
Now it is only used as its name implies, and not to set the overlap window
size as a side-effect. The overlap window in the EdgeAligner is now swept
across a range of sizes from the nominal size up to 10% of the tile size
(this limit could become a tunable parameter if neccessary) in order to find
the size that yields the lowest error. This does incur more computation than
the previous single window size approach, but recent improvements to alignment
should more than make up for it.