Fix defect in mosaic tile blending logic
The tile blending uses an alpha mask computed from the distance transform, but
was not taking into account the possibility of zero-valued pixels in the actual
images. A morphological hole-removal step was added to fix this problem.
Ultimately this should be addressed by using the known tile bounds to
initialize the mask for the distance transform, but that will require some
refactoring.