Account for non-integer relative offsets in LayerAligner (#142)
In LayerAligner, when generating tile crops based on the coarse alignment
offset, the relative offsets are rounded to integers in order to use them as
array indices and the fractional portion is discarded. The discarded fractional
component must be added back in to alignment shifts derived from these crops.
This issue went unnoticed for a long time because the fractional component is
zero in many typical cases.