Improve thumbnail quality to aid rotation detection
Thumbnails now have a minimum size of 1000 pixels on the longest side (unless
the actual estimated final mosaic size is smaller). Also cross-blending is used
to combine the tiles in the thumbnail instead of max-intensity projection.
These changes produce a thumbnail that yields a more reliable estimate of any
rotation between cycles.