Improve coarse-alignment thumbnail quality (#180)
Thumbnails now have a minimum size of 1000 pixels on the longest side and
cross-blending is used to combine the tiles in the thumbnail instead of
max-intensity projection. This improves coarse alignment for very small images.
It also yields a more reliable estimate of rotation between cycles (correction
of which is coming soon).