Correct handling of linear model fitting failure
Now we look at the determinant of the matrix of coefficients, not just
whether all the individual values are zero. Also the intercept is now
zeroed out which fixes the single tile placed far from the rest
(even though this was a sign of complete EdgeAligner failure). A
warning is emitted in the failure case to alert the user of the problem.